backend/pyproject.toml
[project]
name = "porta-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosqlite>=0.21.0",
    "ddgs>=9.5.5",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.116.1",
    "ipykernel>=6.30.1",
    "jinja2>=3.1.6",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.19",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "langsmith>=0.4.25",
    "numpy>=2.2.6",
    "pandas>=2.3.2",
    "pydantic-settings>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.43",
    "uvicorn>=0.35.0",
    "yfinance>=0.2.65",
]

